LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/nested-declarations-cssom-whitespace-expected.txt


FAIL Empty CSSNestedDeclarations do not affect outer serialization assert_equals: expected ".a {\n  & { }\n  & { }\n}" but got ".a {\n  & { }\n  \n  & { }\n  \n}"
FAIL Empty CSSNestedDeclarations do not affect outer serialization (nested grouping rule) assert_equals: expected "@media (width > 1px) {\n  & { }\n  & { }\n}" but got "@media (width > 1px) {\n  & { }\n  \n  & { }\n  \n}"
